made news page in sync with db

+++ application/controllers/Api.php
@@ -101,11 +101,11 @@ class Api extends CI_Controller {
         $config['last_tag_open'] = '<li><a onClick="' . $function . '">';
         $config['last_tag_close'] = '</a></li>';
 
-        $function = "filter_blog('" . ($para1 - $config['per_page']) . "')";
+        $function = "filter_blog('" . (int) ($para1 - $config['per_page']) . "')";
         $config['prev_tag_open'] = '<li><a onClick="' . $function . '">';
         $config['prev_tag_close'] = '</a></li>';
 
-        $function = "filter_blog('" . ($para1 + $config['per_page']) . "')";
+        $function = "filter_blog('" . (int) ($para1 + $config['per_page']) . "')";
         $config['next_link'] = '&rsaquo;';
         $config['next_tag_open'] = '<li><a onClick="' . $function . '">';
         $config['next_tag_close'] = '</a></li>';
@@ -297,13 +297,14 @@ class Api extends CI_Controller {
         $config['last_tag_open'] = '<li><a onClick="' . $function . '">';
         $config['last_tag_close'] = '</a></li>';
 
-        $function = "filter_news('" . ($para1 - $config['per_page']) . "')";
+        $function = "filter_news('" . (int) ($para1 - $config['per_page']) . "')";
         $config['prev_tag_open'] = '<li><a onClick="' . $function . '">';
         $config['prev_tag_close'] = '</a></li>';
 
-        $function = "filter_news('" . ($para1 + $config['per_page']) . "')";
+        $function = "filter_news('" . (int) ($para1 + $config['per_page']) . "')";
         $config['next_link'] = '&rsaquo;';
         $config['next_tag_open'] = '<li><a onClick="' . $function . '">';
+    
         $config['next_tag_close'] = '</a></li>';
 
         $config['full_tag_open'] = '<ul class="pagination">';
@@ -348,6 +349,7 @@ class Api extends CI_Controller {
 
 
 }
+
